refactor: чтение и разбор температур вынес в отдельный клсс
Таким образом можно будет реализовать одинаковое чтение с разных сайтов.
