+++ ImportTemperatureMeteoInfo/ImportTemperatureMeteoInfo.csproj
@@ -7,10 +7,15 @@
     <Company>ООО "ЛЭРС УЧЁТ"</Company>
     <Copyright>Copyright ©  2017 - 2020, ООО ЛЭРС УЧЁТ</Copyright>
     <Product>Система диспетчеризации ЛЭРС УЧЁТ</Product>
+    <ApplicationIcon />
+    <OutputType>Exe</OutputType>
+    <StartupObject />
+    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="CommandLineParser" Version="2.7.82" />
+    <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="3.1.0" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <Reference Include="Lers.System" />
     <Reference Include="System.Net.Http" />
