--- ImportTemperatureMeteoInfo/packages.config
@@ -1,34 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="CommandLineParser" version="2.6.0" targetFramework="net46" requireReinstallation="true" />
-  <package id="Microsoft.AspNetCore.Connections.Abstractions" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.AspNetCore.Http.Connections.Client" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.AspNetCore.Http.Connections.Common" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.AspNetCore.Http.Features" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.AspNetCore.SignalR.Client" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.AspNetCore.SignalR.Client.Core" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.AspNetCore.SignalR.Common" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.AspNetCore.SignalR.Protocols.Json" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="1.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.Configuration" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.Configuration.Abstractions" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.Configuration.Binder" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.Logging" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.Logging.Abstractions" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.Options" version="3.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.Primitives" version="3.1.0" targetFramework="net48" />
-  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net46" />
-  <package id="System.Buffers" version="4.5.0" targetFramework="net48" />
-  <package id="System.ComponentModel.Annotations" version="4.7.0" targetFramework="net48" />
-  <package id="System.IO.Pipelines" version="4.7.0" targetFramework="net48" />
-  <package id="System.Memory" version="4.5.3" targetFramework="net48" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net48" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.0" targetFramework="net48" />
-  <package id="System.Text.Encodings.Web" version="4.7.0" targetFramework="net48" />
-  <package id="System.Text.Json" version="4.7.0" targetFramework="net48" />
-  <package id="System.Threading.Channels" version="4.7.0" targetFramework="net48" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.2" targetFramework="net48" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net48" />
-</packages>
